+++ src/composite/approximate_conditioning.jl
