+++ src/composite/product.jl
