+++ src/gp/kernel.jl
