+++ src/gp/mean.jl
