+++ test/abstract_gp.jl
