+++ test/composite/integrate.jl
