+++ test/composite/product.jl
