+++ test/gp/kernel.jl
