+++ test/gp/mean.jl
