--- test/mean_and_kernel/util.jl
@@ -1,3 +1,0 @@
-@testset "util" begin
-    
-end
